refactor: 상수 역할을 하는 enum class 일부를 domain/constants 패키지로 이동
