+++ src/main/java/bridge/domain/constants/Constant.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge.domain.constants;
 
 public enum Constant {
     MINIMUM_LENGTH_INCLUSIVE(3),
