+++ src/main/java/bridge/domain/constants/GameState.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge.domain.constants;
 
 public enum GameState {
     RUNNING, PAUSE, CLEAR, FAIL;
